logBack file created

+++ src/main/java/com/aster/Main/Service/IMPL/CartServiceImpl.java
@@ -1,5 +1,6 @@
 package com.aster.Main.Service.IMPL;
 
+import com.aster.Main.DTO.CartDto;
 import com.aster.Main.Entity.Cart;
 import com.aster.Main.Entity.CartEntry;
 import com.aster.Main.Entity.Product;
@@ -9,12 +10,14 @@ import com.aster.Main.Repository.CartRepository;
 import com.aster.Main.Repository.ProductRepository;
 import com.aster.Main.Repository.UserRepository;
 import com.aster.Main.Service.CartService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
 import java.util.Set;
 @Service
+@Slf4j
 public class CartServiceImpl implements CartService {
     @Autowired
     private CartRepository cartRepository;
@@ -27,16 +30,28 @@ public class CartServiceImpl implements CartService {
 
 
     @Override
-    public Cart addItemToCart(int sku, int quantity, int id) {
-        User user=userRepository.findById(id).get();
+    public Cart createCart(CartDto cartDto) {
+        log.info("new cart is created for user{}",cartDto.getId());
+        Cart cart=Cart.builder()
+                .grandTotal(cartDto.getGrandTotal())
+                .build();
+        cart.setUser(userRepository.findById(cartDto.getId()).get());
+
+        return cartRepository.findById(cart.getCartId()).get();
+    }
+
+    @Override
+    public Cart addItemToCart(int sku, int quantity,int id) {
+        Cart cart=cartRepository.findById(id).get();
+        User user= cart.getUser();
         Product product=productRepository.findById(sku).get();
-        Cart cart=user.getCart();
+
         if(cart==null){
             cart=new Cart();
         }
         Set<CartEntry> cartitems = cart.getCartEntrySet();
         CartEntry cartEntry = findCartItem(cartitems,product.getSku())  ;
-        if (cartitems == null) {
+        if (cartitems == null && product.getQuantity()>quantity) {
             cartitems = new HashSet<>();
             if (cartEntry == null) {
                 cartEntry = new CartEntry();
@@ -76,44 +91,53 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public Cart updateItemInCart(Product product, int quantity,User user) {
-        Cart cart = user.getCart();
+    public Cart updateItemInCart(Product product, int quantity,Cart cart) {
 
         Set<CartEntry> cartItems = cart.getCartEntrySet();
 
         CartEntry cartEntry = findCartItem(cartItems, product.getSku());
+        if(quantity==0){
+            cartItems.remove(cartEntry);
 
-        cartEntry.setQuantity(quantity);
-        cartEntry.setTotalPrice(quantity * product.getPrice());
+            cartEntryRepository.delete(cartEntry);
+        }
+        else {
 
-        cartEntryRepository.save(cartEntry);
+            cartEntry.setQuantity(quantity);
+            cartEntry.setTotalPrice(quantity * product.getPrice());
+
+            cartEntryRepository.save(cartEntry);
+        }
 
         double totalPrice = totalPrice(cartItems);
         cart.setGrandTotal(totalPrice);
 
         return cartRepository.save(cart);
+    }
+
+    @Override
+    public Cart getCart(int id) {
+        return cartRepository.findById(id).get();
     }
     // {user:products:[ product1, product2 ]}
 
-    @Override
-    public Cart deleteItemFromCart(Product product,User user) {
-        Cart cart = user.getCart();
-
-        Set<CartEntry> cartItems = cart.getCartEntrySet();
-
-        CartEntry cartEntry = findCartItem(cartItems, product.getSku());
-
-        cartItems.remove(cartEntry);
-
-        cartEntryRepository.delete(cartEntry);
-
-        double totalPrice = totalPrice(cartItems);
-
-        cart.setCartEntrySet(cartItems);
-        cart.setGrandTotal(totalPrice);
-
-        return cartRepository.save(cart);
-    }
+//    @Override
+//    public Cart deleteItemFromCart(Product product,User user) {
+//        Cart cart = user.getCart();
+//
+//        Set<CartEntry> cartItems = cart.getCartEntrySet();
+//
+//        CartEntry cartEntry = findCartItem(cartItems, product.getSku());
+//
+//
+//
+//        double totalPrice = totalPrice(cartItems);
+//
+//        cart.setCartEntrySet(cartItems);
+//        cart.setGrandTotal(totalPrice);
+//
+//        return cartRepository.save(cart);
+//    }
 
     private CartEntry findCartItem(Set<CartEntry> cartEntrySet, int productId) {
         if (cartEntrySet == null) {
